Add initial tests to character-level interaction
Which were rather successful.
Also added initial logic to flip world layers (effectively mirroring the world), which was also successful. For this, we need that every level be composed of our custom TileMap actor.
We need to remember to configure the level collision profile to use our custom profile, PaperLevel, and configure the PhysicalMaterial to be PM_Level. Maybe we can create a blueprint actor derived from said custom actor with these defaults already configured.

MMN-008